rename tools file name
