+++ src/jobtrendx/email_processor.py
@@ -7,7 +7,7 @@ from pathlib import Path
 
 import email
 
-from . import tools
+from . import tools_processor as tools
 from . import logger
