+++ tests/test_tools.py
@@ -12,7 +12,7 @@ import pytest
 
 import pandas as pd
 
-from jobtrendx.tools import check_directory, check_dir_not_empty, \
+from jobtrendx.tools_processor import check_directory, check_dir_not_empty, \
     returns_all_files_in_dir, returns_eml_files, returns_eml_path, \
     _extract_attachments, _clean_eml_payload, eml_to_dataframe, \
     detect_language, _detect_single_language, _check_language
